feature: extract controller
create new ExpenseCategoriesController
