+++ src/main/java/ml/meajudadev/despesas/v1/ExpenseCategoriesController.java
@@ -1,0 +1,26 @@
+package ml.meajudadev.despesas.v1;
+
+import ml.meajudadev.despesas.ExpenseCategory;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+public class ExpenseCategoriesController {
+    private List<ExpenseCategory> categories;
+
+    public ExpenseCategoriesController() {
+        categories = List.of(
+                new ExpenseCategory(1, 1, "Alimentação", 'A', 'E'),
+                new ExpenseCategory(2, 1, "Saúde", 'A', 'E'),
+                new ExpenseCategory(3, 1, "Entretenimento", 'A', 'E'),
+                new ExpenseCategory(4, 1, "Educação", 'A', 'E')
+        );
+    }
+
+    @GetMapping("/api/v1/expense-categories")
+    public List<ExpenseCategory> listCategories() {
+        return categories;
+    }
+}
